requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
joblib==1.4.2
scikit-learn==1.6.1
numpy==2.2.6
pandas==2.2.3
reportlab==4.2.5
qrcode[pil]==8.0
python-multipart==0.0.12
slowapi==0.1.9
aiofiles==24.1.0